Refactor event handling by extracting helper functions for improved readability and maintainability
