+++ tests/plugins/director/json/json_reporter/handle_events.py
@@ -1,7 +1,4 @@
-from argparse import ArgumentParser, Namespace
-from io import StringIO
-from typing import Any, Optional
-from unittest.mock import Mock, call
+from unittest.mock import call
 
 from vedro import given, scenario, then, when
 from vedro._test_utils import (
@@ -16,74 +13,15 @@ from vedro._test_utils import (
     make_startup_event,
     make_vscenario,
 )
-from vedro.core import Config, ConfigType, Dispatcher
-from vedro.core.exc_info import TracebackFilter
-from vedro.events import ArgParsedEvent, ArgParseEvent, ConfigLoadedEvent
-from vedro.plugins.director import Director, DirectorPlugin
-from vedro.plugins.director.json import JsonFormatter, JsonReporter, JsonReporterPlugin
-from vedro.plugins.director.json._json_reporter import JsonFormatterFactory
 
-
-def make_config() -> ConfigType:
-    class TestConfig(Config):
-        class Registry(Config.Registry):
-            TracebackFilter = lambda modules: Mock(spec_set=TracebackFilter)  # noqa: E731
-    return TestConfig
-
-
-async def fire_config_loaded_event(dispatcher: Dispatcher,
-                                   config: Optional[ConfigType] = None) -> None:
-    if config is None:
-        config = make_config()
-    config_loaded_event = ConfigLoadedEvent(config.path, config)
-    await dispatcher.fire(config_loaded_event)
-
-
-async def fire_arg_parsed_event(dispatcher: Dispatcher) -> None:
-    arg_parse_event = ArgParseEvent(ArgumentParser())
-    await dispatcher.fire(arg_parse_event)
-
-    arg_parsed_event = ArgParsedEvent(Namespace())
-    await dispatcher.fire(arg_parsed_event)
-
-
-def make_json_reporter(formatter_factory: JsonFormatterFactory) -> JsonReporterPlugin:
-    class _JsonReporter(JsonReporter):
-        class RichReporter(JsonReporter.RichReporter):
-            enabled = False
-
-    output = StringIO()
-    reporter = JsonReporterPlugin(_JsonReporter, formatter_factory=formatter_factory,
-                                  output=output)
-    return reporter
-
-
-def make_director_plugin(dispatcher: Dispatcher) -> DirectorPlugin:
-    class _Director(Director):
-        default_reporters = ["json"]
-
-    director = DirectorPlugin(_Director)
-    director.subscribe(dispatcher)
-
-    return director
-
-
-def make_formatter_mock(**kwargs: Any) -> JsonFormatter:
-    return Mock(spec_set=JsonFormatter, **kwargs)
-
-
-def make_formatted_event_mock(event_name: str = "<event>") -> Mock:
-    return Mock(return_value={"event": event_name})
+from ._helpers import fire_arg_parsed_event, fire_config_loaded_event, setup_json_reporter
 
 
 @scenario
 async def trigger_startup_event():
     with given:
-        make_director_plugin(dispatcher := make_dispatcher())
-
-        formatter_ = make_formatter_mock(format_startup_event=make_formatted_event_mock())
-        reporter = make_json_reporter(lambda tb_filter: formatter_)
-        reporter.subscribe(dispatcher)
+        dispatcher = make_dispatcher()
+        reporter, formatter_ = setup_json_reporter(dispatcher, mocked_event="startup")
 
         await fire_config_loaded_event(dispatcher)
         await fire_arg_parsed_event(dispatcher)
@@ -102,11 +40,8 @@ async def trigger_startup_event():
 @scenario
 async def trigger_startup_event_with_scheduled():
     with given:
-        make_director_plugin(dispatcher := make_dispatcher())
-
-        formatter_ = make_formatter_mock(format_startup_event=make_formatted_event_mock())
-        reporter = make_json_reporter(lambda tb_filter: formatter_)
-        reporter.subscribe(dispatcher)
+        dispatcher = make_dispatcher()
+        reporter, formatter_ = setup_json_reporter(dispatcher, mocked_event="startup")
 
         await fire_config_loaded_event(dispatcher)
         await fire_arg_parsed_event(dispatcher)
@@ -130,11 +65,8 @@ async def trigger_startup_event_with_scheduled():
 @scenario
 async def trigger_scenario_run_event():
     with given:
-        make_director_plugin(dispatcher := make_dispatcher())
-
-        formatter_ = make_formatter_mock(format_scenario_run_event=make_formatted_event_mock())
-        reporter = make_json_reporter(lambda tb_filter: formatter_)
-        reporter.subscribe(dispatcher)
+        dispatcher = make_dispatcher()
+        reporter, formatter_ = setup_json_reporter(dispatcher, mocked_event="scenario_run")
 
         await fire_config_loaded_event(dispatcher)
         await fire_arg_parsed_event(dispatcher)
@@ -153,11 +85,8 @@ async def trigger_scenario_run_event():
 @scenario
 async def trigger_scenario_passed_event():
     with given:
-        make_director_plugin(dispatcher := make_dispatcher())
-
-        formatter_ = make_formatter_mock(format_scenario_passed_event=make_formatted_event_mock())
-        reporter = make_json_reporter(lambda tb_filter: formatter_)
-        reporter.subscribe(dispatcher)
+        dispatcher = make_dispatcher()
+        reporter, formatter_ = setup_json_reporter(dispatcher, mocked_event="scenario_passed")
 
         await fire_config_loaded_event(dispatcher)
         await fire_arg_parsed_event(dispatcher)
@@ -176,11 +105,8 @@ async def trigger_scenario_passed_event():
 @scenario
 async def trigger_scenario_failed_event():
     with given:
-        make_director_plugin(dispatcher := make_dispatcher())
-
-        formatter_ = make_formatter_mock(format_scenario_failed_event=make_formatted_event_mock())
-        reporter = make_json_reporter(lambda tb_filter: formatter_)
-        reporter.subscribe(dispatcher)
+        dispatcher = make_dispatcher()
+        reporter, formatter_ = setup_json_reporter(dispatcher, mocked_event="scenario_failed")
 
         await fire_config_loaded_event(dispatcher)
         await fire_arg_parsed_event(dispatcher)
@@ -199,11 +125,8 @@ async def trigger_scenario_failed_event():
 @scenario
 async def trigger_scenario_skipped_event():
     with given:
-        make_director_plugin(dispatcher := make_dispatcher())
-
-        formatter_ = make_formatter_mock(format_scenario_skipped_event=make_formatted_event_mock())
-        reporter = make_json_reporter(lambda tb_filter: formatter_)
-        reporter.subscribe(dispatcher)
+        dispatcher = make_dispatcher()
+        reporter, formatter_ = setup_json_reporter(dispatcher, mocked_event="scenario_skipped")
 
         await fire_config_loaded_event(dispatcher)
         await fire_arg_parsed_event(dispatcher)
@@ -222,13 +145,8 @@ async def trigger_scenario_skipped_event():
 @scenario
 async def trigger_scenario_reported_event():
     with given:
-        make_director_plugin(dispatcher := make_dispatcher())
-
-        formatter_ = make_formatter_mock(
-            format_scenario_reported_event=make_formatted_event_mock()
-        )
-        reporter = make_json_reporter(lambda tb_filter: formatter_)
-        reporter.subscribe(dispatcher)
+        dispatcher = make_dispatcher()
+        reporter, formatter_ = setup_json_reporter(dispatcher, mocked_event="scenario_reported")
 
         await fire_config_loaded_event(dispatcher)
         await fire_arg_parsed_event(dispatcher)
@@ -248,11 +166,8 @@ async def trigger_scenario_reported_event():
 @scenario
 async def trigger_cleanup_event():
     with given:
-        make_director_plugin(dispatcher := make_dispatcher())
-
-        formatter_ = make_formatter_mock(format_cleanup_event=make_formatted_event_mock())
-        reporter = make_json_reporter(lambda tb_filter: formatter_)
-        reporter.subscribe(dispatcher)
+        dispatcher = make_dispatcher()
+        reporter, formatter_ = setup_json_reporter(dispatcher, mocked_event="cleanup")
 
         await fire_config_loaded_event(dispatcher)
         await fire_arg_parsed_event(dispatcher)
